Split AbstractSelectionBehaviour and AbstractDragSelectionBehaviour into two classes
* This is to prepare other selection behaviours than drag selection behaviours as well
