+++ src/main/java/org/mastodon/grapher/opengl/overlays/AbstractSelectionBehaviour.java
@@ -18,21 +18,19 @@ public abstract class AbstractSelectionBehaviour extends AbstractNamedBehaviour 
 {
 	protected final SelectionModel< Spot, Link > selection;
 
-	private final ScreenTransformState screenTransformState;
+	protected final ScreenTransformState screenTransformState;
 
 	protected final ScreenTransform screenTransform;
 
 	protected final PointCloudPanel pointCloudPanel;
 
-	private final ReentrantReadWriteLock lock;
+	protected final ReentrantReadWriteLock lock;
 
 	protected final ModelGraph graph;
 
 	protected final FocusModel< Spot > focus;
 
-	private final boolean addToSelection;
-
-	protected boolean dragging = false;
+	protected final boolean addToSelection;
 
 	AbstractSelectionBehaviour( final String name, final SelectionModel< Spot, Link > selection, final FocusModel<Spot> focus,
 			final ModelGraph graph, final PointCloudPanel pointCloudPanel,
@@ -47,59 +45,6 @@ public abstract class AbstractSelectionBehaviour extends AbstractNamedBehaviour 
 		this.addToSelection = addToSelection;
 		this.screenTransformState = pointCloudPanel.getScreenTransform();
 		this.screenTransform = new ScreenTransform();
-		pointCloudPanel.getCanvas().overlays().add( this );
-	}
-
-	protected abstract void doDrag( final int x, final int y );
-
-	protected abstract void doInit( final int x, final int y );
-
-	protected void doEnd(final int x, final int y)
-	{
-		// do nothing
-	}
-
-	@Override
-	public void init( final int x, final int y )
-	{
-		screenTransformState.get( screenTransform );
-		doInit( x, y );
-		dragging = false;
-		pointCloudPanel.overlayChanged();
-	}
-
-	@Override
-	public void drag( final int x, final int y )
-	{
-		screenTransformState.get( screenTransform );
-		doDrag( x, y );
-		if ( !dragging )
-			dragging = true;
-		pointCloudPanel.overlayChanged();
-	}
-
-	@Override
-	public void end( final int x, final int y )
-	{
-		doEnd( x, y );
-		if ( dragging )
-		{
-			dragging = false;
-			lock.readLock().lock();
-			try
-			{
-				select();
-			}
-			finally
-			{
-				lock.readLock().unlock();
-			}
-		}
-		else
-		{
-			selection.clearSelection();
-		}
-		pointCloudPanel.overlayChanged();
 	}
 
 	@Override
